examples/smolagents/requirements.txt
duckduckgo_search == 8.1.1
gradio == 5.49.1
openai == 2.6.1
pandas == 2.3.3
python-dotenv == 1.2.1
requests == 2.32.5
smolagents == 1.22.0
